readme: remove unneeded symlink from build instructions
Previously, our module was located under `modules/$MODULE` in
the west workspace. However, with the newer flow using the
west manifest from the module itself, the path of the module
is hard-coded by west to use `$MODULE.git` instead.

In order to keep our flow the same, we made a symlink, but
it's unnecessary to do so. Instead, just use the
`$MODULE.git` path.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>