boards: tenstorrent: update openocd setup to run to bootrom completion
Update the reset handling code for openocd to run to the completion
point of the bootrom. We achieve this by setting a breakpoint at the end
of bootrom execution (just before branching to the application image)

Since the bootrom is now running to completion before we attempt to load
an image to RAM, PLL1 (used for ARCCLK) will be initialized, and openocd
can load to CSM correctly. This means that `west flash` can write an
image to RAM and execute it. We can also drop a bunch of openocd reset
hooks intended to emulate the behavior of the bootrom.

This change was verified by changing telemetry data in the CMFW
application, flashing the new CMFW binary with `west flash`, and
verifying that the card reappeared on the PCIe bus, and that the new
telemetry data was reflected in `tt-smi`

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>