kconfig: housekeeping so that Kconfig compliance check passes
Previously, we had a few errors in the Kconfig CheckCompliance
module when run as

./zephyr/scripts/ci/check_compliance.py --annotate \
  -c $(git rev-parse HEAD)..

this was partially because the upstream check was only used to
check the main Zephyr repository. However, it would be useful
to reuse the Kconfig compliance check for other modules too.

Running the check identified some Kconfig values that were
not defined, which prompted some cleanup.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>