scripts: update restore-stable-bmfw to handle multiple JLinks
Update restore-stable-bmfw script to handle multiple JLink devices being
connected to the host CI machine. We accomplish this by adding a script,
"get-board-id", which gets the "id" parameter from a hardware map file
in order to pass to "west flash". This way we target the correct JLink
device when restoring our firmware image

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>