drivers: serial: add tenstorrent virtual pcie uart driver
This virtual serial driver manages two self-describing
ring-buffers / fifos mapped from Blackhole CSM over PCIe.

One fifo is for TX, and the other is for RX. Both are
from the perspective of the device (so ARC in this case).

Currently, the driver only works in polling mode, although
we can look into enabling MSI-based interrupts at a later
time.

Important to note is that our uart_poll_out() method will
not actually block when the FIFO is full, since that could
potentially hang the system if no host-side process is
emptying the FIFO. So we offer a choice to drop, overwrite
the oldest (head) or the newest (tail) character.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>