drivers: smbus: tt_smbus: enforce read block size restriction
The Zephyr SMBUS API defines that the maximum block read size is 32
bytes. Previously, we did not check to make sure we would not exceed
this boundary when reading from the SMBUS device. Add a check for this,
and provide a full 32 byte buffer when calling smbus_block_read()

This resolves an issue with the BMC firmware that could be triggered
when resetting the SMC while the BMC was polling for SMBUS data. The SMC
would generate invalid data on the I2C lines while resetting, and the
smbus_block_read() function would read 255 bytes into the provided
buffer, causing a stack overflow.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>