tenstorrent: jtag_bootrom: do not refer to non-existent kconfig
We need to enable the Kconfig compliance checker, and these few
lines are preventing that from happening at the moment.

I believe that the characterization boards are not meant to be
listed in the Open Source repository, so we should remove refs
to undefined Kconfig options.

However, it would be good to preserve the correct behaviour in
the case we would like to have an internal zephyr module for
characterization boards, so let's try to keep the functionality
and still pass compliance.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>