manifest: bump zephyr revision
Update our module's manifest (`west.yml`) to point at revision
769df0b579b09ee96592c268163300ccffb87a8f of Zephyr, as we
approach 4.1.0-rc1 which is scheduled to be released in 2 weeks.

We should be able to bump this revision with the expectation
that things won't break on a fairly regular basis, and doing so
ensures that we actively engage in patch management (i.e. drop
or rework patches as needed that are listed in
zephyr/patches.yml).

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>