tt_boot_fs: schema, script, and tests
tt_boot_fs image generation isn't done, but this is in a safe
enough place to check it in for others to build on. I would like
to mostly use existing code for this (e.g. boot_fs.py, and
generate_boot_fs.py, tt_boot_fs.c, etc). However, I also took
the liberty to create a pykwalify schema for tt_boot_fs specs
and also to introduce a few new changes:

1. use a common schema for images and fail_over_image
2. use a dedicated name field for image names, instead of key
3. add the padto property for fixed-size partitions

The 2nd point basically highlights a quirk (maybe shortcoming?)
in pykwalify, in that key names for maps must be known in
advance, but it allows for slightly more conventional
verification of specs and processing of images.

The 'source' field allows us to specify a start address of
images in tt_boot_fs (i.e. the offset from start of flash).

W.r.t. the 3rd point, 'padto' allows us to specify a size to pad
the image to. The combination of 'source' and 'padto' allows us
to create fixed-size partitions that tt_boot_fs will also
understand. Note: we may want to add a 'padbyte' property to
the 'alignment' object to specify what byte value should be used
for padding (normally this is 0xff for flash, but for emulation
devices, it can be 0x00).

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>