tests: tenstorrent: convert bitrev to a unit test
For tests that are architecture and hardware independent and are
restrcted to memory and compute, Zephyr supports a special
"unit" test type classification.

No cross compilation happens, nor do these tests run in any kind
of emulator or simulator; they are just host-compiled binaries.

Unit-type tests tend to build and run faster than all other
type of test in Zephyr.

The original reason that this test was not added as a unit type
test was so that it could be tested easily on macOS.

Since this test does not need to be run on a regular basis, and
the implementation has rarely ever changed, it's safe to move
this over. It will be run in CI (on Linux) regularly.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>