ci: use alt-config-root for in tree test cases
Use twister's alt-config-root argument for in tree test cases. This
allows us to define test YAMLs out of tree for in tree tests- which
enables us to drop a patch to the Zephyr tree. As part of this, tag
tests we intend to run with "smoke". This way, we don't have to edit the
CI invocation for every new test we want to run on hardware

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>