boards: tenstorrent: read PC directly from ARC, remove steps
It seems that the step commands within the reset workaround cause more
harm than good, as the core is sometimes reset (and in a spin loop) by
the time these commands run, which results in a failure to step.

Remove the step commands and read the PC from the ARC AUX registers,
which bypasses the register cache within openocd.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>