tenstorrent: bh_chip: add pgood fault interrupt
Adds interrupts on pgood gpio fall and rise to enable the pgood fault sequence.

When a pgood fault is detected, the board fault LED is turned on and ASIC reset
is asserted. If pgood comes back up, then the LED is cleared and the ASIC is
brought out of reset. If it falls again within 1 second, the pgood severe fault
state is entered and can only be cleared with a power cycle.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>