app: smc: avoid busy-looping in main()
This busy-loop was added as a workaround to ARC hanging once
`main()` returns. A root cause still needs to be found for that.

The main problem with `k_yield()` is that it is effectively a
busy loop when no other threads are ready to run, or when other
threads have a lower (numerically higher) priority.

Busy looping is generally bad. We want the scheduler to be
capable of scheduling other threads to run. The simplest way to
do this is to explicitly unschedule the thread for some
duration, which was arbitrarily chosen as 1s in this case.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>