bmc: Updated flash config to allow non-volatile register config to apply
There is an issue on the p300 boards where the spi flash is held in
hold when the bmfw attempt to access it. To work around this I modified
the non volatile registers of the eeprom flash to disable the hold pin.
It appears that this changes the behaviour of the eeprom slightly where
it takes slightly longer to apply this change. Having the spi-nor driver
reset the device during init appears to give the time to allow this
change to apply.

Also disabled chip1_spi and made chip0 the primary because the current p300
board has another spi issue where the clk pin is not mapped to hardware.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>