boards: pad bmfw to nearest 2048 byte boundary
Due to a bug in the update code for BMFW 0.3.1, the erase size is determined by
the size of the image rounded to the nearest write size boundary. On the
STM32 based BMC, the write size != the erase size, so erases will fail
based on this rounding (and updates will not work)

To work around this, pad the BMFW image to a 2048 byte boundary in the
tt_boot_fs. This will enable updates to work, at the cost of a bit more
flash space used

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>