ci: update SMC to use tt-console by default
Update SMC to use the virtual PCIe uart / tt-console by default
everywhere.

This is a 3-part change. The first part being updating all of the
tests and default configuration to enable CONFIG_TT_VUART and
to point to `/opt/tenstorrent/twister/hw-map-vuart.yml`.

The second part will be copying hw-map-vuart.yml to hw-map.yml
on CI machines.

The 3rd part of this change will be to change hw-map-vuart.yml to
hw-map.yml in the repository.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>