tests: boot: mcuboot: skip mcuboot test since it is covered
The zephyr testsuite `tests/boot/test_mcuboot` is poorly
conditioned in that if mcuboot already exists in the
bootloader partition, it can derail the test entirely.

The ideal way to structure the test would be to
* erase the bootloader partition
* erase the slot0 partition
* write to the slot1 partition
* write to the slot0 partition
* write to the bootloader partition

Otherwise, between writes, if a reset occurs (which is common)
the bootloader can already jump to whatever application occupies
slot0, etc.

Additionally, we already cover the intended end-user flashing
procedure in `app/smc/app.e2e-smoke-flash` and
`app/smc/app.recovery-flash`.

The mcuboot test was useful for the original procedure though.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>