scripts: tt_boot_fs.py: add ls subcommand to list fs contents
Add a ls (list) subcommand to list all of the "files" contained
in a given tt_boot_fs.

This should make it significantly easier to examine these files
boot fs contents by both humans and machines.

The '-q' or '--quiet' option skips printing the title row
and separator, which should support command line processing
better. If specified twice, no output is printed at all, which
can be useful when used as a python library.

The '-j' or '--json' option prints the listing to JSON.

The '-b' or '--base64' option assumes that the input is base64
encoded (i.e. text) rather than raw binary, which is the case
for image.bin files in fwbundle packs, while the immediate
raw build output tt_boot_fs.bin is just binary. This means
we can list both base64 and binary encoded boot fs images.

I have not attempted to first convert from hex, but that
could be added in a subsequent PR.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>