scripts: add a script for CI to check if a runner exists with a label
Add scripts/check-runners.py which will check if any available
GitHub runner exists with a specific label.

The script loads a list of potential runners from .github/runners.yml .

It prints the output either in JSON or grep'able text in the format
"board_rev label".

The resulting JSON can be used in GitHub Actions YAML to generate a
job matrix.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>