board: tt_blackhole: minor cleanup of SRST workaround
Add a minor cleanup of the SRST workaround, bringing variables out to
Kconfig. It makes the SRST workaround optional, but enabled by default.

By doing so, we preserve the previous functionality, which will also
work when only one JTAG adapter is present for SMC. This will work
unmodified for P100, but requires the SRST (BH_RESET) signal is brought
out to the debug header for other Blackhole cards such as P100A, as
well as P150 and P300 variants.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>