boards: tt_blackhole: only write vars.cfg when kconfig exists
When building the app/smc target with --sysbuild, then output
file is written 3 times. At least the last of those times,
CONFIG_TT_SRST_ASSIST is unset, which only means the first
line is written to the file.

Only write bmc-reset-vars.cfg when CONFIG_TT_SRST_ASSIST is
defined so that the file is fully populated.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>