maintainers: add release notes maintainer to the maintainers file
Add a release notes maintainer to the maintainers file.

This file intentionally uses the same structure as that from the Zephyr
Project. It is intended to be used programmatically to associate areas
of code (C, devicetree, documentation, etc) to various individuals
within Tenstorrent's firmware organization.

Additional areas and maintainers should be added soon.

Please see the upstream maintainers file for more info:

https://github.com/zephyrproject-rtos/zephyr/tree/main/MAINTAINERS.yml

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>