app: dmc: add smbus testsuite for E2E testing
Add an smbus testsuite for E2E testing. This testsuite requires several
changes, because we need the DMC application to act as an SMBUS
controller for the SMC.

We add a test routine that can be enabled with a Kconfig, and at the
exit of this test routine we write a pass result to the scratch register
of the SMC. In the E2E test, we validate the value of this scratch
register to confirm the SMBUS tests passed as expected.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>