drivers: i2c + smbus: remove forked stm32 drivers
The forked stm32 i2c and smbus drivers should not be carried in-tree as
they are effectively technical debt.

Convert them to patches that can more easily be upstreamed by the
original author.

Ideally, there would be nothing left to keep in-tree, but if there is,
it should be extremely minimal.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>