scripts: rtt: adjust permissions of rtt scripts
Adjust the permissions of the rtt scripts dmc_rtt.py and smc_rtt.py
to be executable since they have a shebang.

Remove the shebang from rtt_helper.py since it does not have a main
function and is mainly intended to be used as an import source by
python code.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>