bh_arc: add support for tensix reset message
Tensix resets are needed for internal testing purposes. CMFW needs to
know when the Tensix reset is happening, since this also resets the NOC.
If NOC traffic from CMFW is in flight while NOC gets reset, this can
cause CMFW's ARC processor to hang. Having CMFW do the Tensix reset is
the simplest way of solving this.

Signed-off-by: Aaron Fong <afong@tenstorrent.com>