scripts: pci_utils: support removing and rescanning the bus via sudo
In certain CI jobs that are run inside of a privileged container,
processes run as root so there are never any permission issues.

As a user e.g. on a lab machine, when attempting to reproduce CI
failures, we don't have the same permission, but we do want the scripts
to succeed anyway.

This is particularly true when trying reproduce a failure with twister
the usual way (by copying and modifying the repro-recipe that twister
prints).

Echo bits to the `remove` and `rescan` nodes as root via sudo.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>