dmc: add handling for auto reset timeout and heartbeat messages
Add a heartbeat and auto reset timeout messages to the DMC firmware.
The SMC peripheral can request a reset timeout with the auto reset
timeout message, and then must send heartbeat pings within this
interval. If the SMC fails to ping, the DMFW will read the current ARC
program counter, record this, and reset the BH ASIC.

Co-authored-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>
Signed-off-by: Petra Alexson <palexson@tenstorrent.com>