app: smc: fall-back to null if wdt0 is disabled
Previouosly, `app/smc` would unconditionally try to get a DT handle on wdt0.
This would fail, however, when the watchdog timer was not enabled in DT.

Use `DEVICE_DT_GET_OR_NULL()` so that we are not left with an undefined
symbol. i.e. get rid of the error that says

```
error: '__device_dts_ord_9' undeclared here
```

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>