doc: release: process: minor correction on tag push
Previously, `git push --tags` was in the release process doc. However that
can possibly have side-effects because it pushes all local tags. We want
to restrict the tag that we push to only the newly created one. For that
purpose, use `git push git@github.com:tenstorrent/tt-firmware.git <tag>`.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>