maintainers: add areas of interest for tenstorrent
Within the MAINTAINERS.yml file, there are or will be several areas of
interest to Tenstorrent in terms of code ownership, delegation of
responsibility for improvement and bugfixes, testing, and so on.

Add each of those areas with a "Maintainer needed!" comment.

This coincides with a salient point brought up at the recent
SysEng F2F meeting where we established that code-ownership is critical
to platform stability and management of technical debt.

Being a maintainer or collaborator is a merit-based role. Before anyone
is added to the MAINTAINERS.yml file, they must demonstrate that they
have made contributions to the area, understand what it takes to make
improvements in the area, and commit to responding to issues in a
timely manner.

The MAINTAINERS.yml file is also used by the assigner.yml workflow. So
anytime someone makes a pull request targeting a specific area, the
assigner can look through the affected files and automatically find the
correct person to review code. Similarly, it also semi-automates the
delegation of issues and bug reports to the correct person.

For more information, please see

https://docs.zephyrproject.org/latest/project/project_roles.html

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>