ci: concurrency: ensure key is unique per job
For the build-fw reusable workflow, we observed the second job canceling
the first job. This was because the `concurrency` key was not unique for
each job in the workflow.

Add `${{ github.job }}` as another item to key from, so that two jobs
within the same workflow do not cancel each other.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>