Updated tt-console to use the kmd tlb allocation API
As of kmd 1.34+ tt-kmd provides an API for allocating TLBs instead of
manually mapping the PCIe bar space. This gives us the ability to run
alongside other programs that wish to also interact with the hardware.

This patch adds support for this capability and removes support for the
older manually mapped TLBs.

Signed-off-by: Daniel Rosen <drosen@tenstorrent.com>