boards: tenstorrent: tt_blackhole: enable mcuboot as opt-in build
Enable mcuboot as an opt-in build feature. This can be enabled by
building the app with -DFILE_SUFFIX=mcuboot. MCUBoot flash partitions
are defined as follows:
- bootloader partition. This is the partition where mcuboot resides. It
  is defined for the build to function, but the true partition offset is
  set within the bootfs yaml descriptor. The mcuboot partition is used
  as the "failover" image by the bootloader, which boots the image using
  the recovery image descriptor.
- image 0 partition. This partition flash location matches where the
  image will be placed in flash by the bootfs. This image is the
  "mission mode" image, which will be upgraded each time tt-flash loads
  a new firmware
- image 1 partition. This partiton flash location matches where the
  image will be placed in flash by the bootfs. This image is the
  recovery mode image, and will not be updated during a normal flash

The recovery image is versioned as 0.0.0, so it will only boot if the
mission mode image is not valid. Both the recovery image and mission
mode image have binary trailer suffixes, described below:

Add magic_confirmed and magic_test binary files. These are binary blobs
that contain the image trailers MCUBoot expects to see, with the
following significance:
- magic_confirmed: marks the image as confirmed (IE tested and
  bootable)
- magic_test: Marks the image as valid and ready for testing. Image
  can be booted, but must confirm itself after booting.

These binary files are placed as suffixes to the failover and smc
mission mode images. The confirmed binary is used for the failover
image, so that the failover image is always considered valid and
bootable. The smc mission mode image uses the test binary, so that the
image will be booted in test mode and must confirm itself.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>