boards: tenstorrent: tt_blackhole: define MCUBoot partitions
Define MCUBoot flash partitions for the tt_blackhole board. These
partitions must be mirrored in the bootfs, so that the CMFW and failover
images are placed where we expect them to be in flash.

Add magic_confirmed and magic_test binary files. These are binary blobs
that contain the image trailers MCUBoot expects to see, with the
following significance:
- magic_confirmed: marks the image as confirmed (IE tested and
  bootable)
- magic_test: Marks the image as valid and ready for testing. Image
  can be booted, but must confirm itself after booting.

These binary files are placed as suffixes to the failover and smc
mission mode images. The confirmed binary is used for the failover
image, so that the failover image is always considered valid and
bootable. The smc mission mode image uses the test binary, so that the
image will be booted in test mode and must confirm itself.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>