app: smc: turn on mcuboot support for SMC application
Turn on MCUBoot support for the SMC application. This will enable
building mcuboot as the "cmfw" image, which in turn will boot the
mission mode image in test mode.

The mission mode image must confirm itself on the first boot, or it will
be removed on the next boot and mcuboot will fall back to the recovery
image.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>