doc: release: shorten release-notes to use major.minor
When we create a patch release, e.g. 18.5.1, we do not want to create
an entirely separate copy of the release notes for the patch release. We
want to update the existing release notes document for that release.

However, the release notes document on a given release was previously
titled e.g. release-notes-18.5.0.md . It would seem counterintuitive to
add 18.5.1 release notes to the 18.5.0 release notes doc.

Therefore, we should do what upstream does as well, and name our
release notes and migration guide documents only with major.minor
instead of major.minor.patch .

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>