ci: add boards.json file for one source of truth of built boards
Add `.github/boards.json` as a single source of truth that controls the
boards that are attempted to be built and tested in CI.

Note: a later PR might add the capability to query which runners are
available in the local github organization or repository. Until that time,
CI will attempt to run HW in the loop tests even when the runner is not
available.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>