ci: bump tt-flash rev to 3.3.5, tt-smi rev to 3.0.23
Smoke tests and soak tests were both failing on p150a during the "Flash
the firmware" step with the error that there were flash region mismatches.
The workflows use version 3.3.3.

http://bit.ly/45WZ7Rs
http://bit.ly/46eEWyL

However, manual repro on the same p150a machine using version 3.3.5 does
not produce any flash region mismatches.

https://pastebin.com/vknhTuZD

Bump the version of the tt-flash to v3.3.5 and the version of tt-smi to
v3.0.23.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>