lib: bh_arc: reduce scratchpad size from 64kB to 36kB
We had some difficulty building our application with debug information
before because the linked binary would not fit into memory reserved for
cmfw.

Reduce the large scratchpad size down from 64 kB to 36kB, which is the
size of ERISC firmware. The size can be further reduced by adjusting the
firmware loading algorithms to be flash-aware. If using malloc is a
possibility then the buffers can be freed when they are no longer
required, which would allow for additional runtime memory usage.

Technically, I think we should also be able to extend the boundary within
CSM (SRAM) where CMFW may be loaded, but if we can get away with using a
smaller static buffer and still meet the same performance and functional
requirements, we might as well do that too.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>