app: smc: add overlays to enable tracing
Add overlays to enable tracing support in the SMC application. We
utilize the VUART as a tracing backend in synchronous mode, since the
backend is very low latency and allows for fast memory transfers.

Note we also correct the system clock hardware cycles per second here so
that the timestamps in event logs will be correct.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>