scripts: tooling: add tt-tracing tool
Add tt-tracing tool, which uses the vuart to trace data. This is
different than the tt-console tooling, because we do not intend to use
the console interactively.

Instead, the program writes "enable\r" to the console to activate
tracing, reads trace data via the console, and writes "disable\r" at
exit to complete tracing.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>