ci: do not retry if twister jobs fail
Historically, `twister` has required the `--retry-failed` argument and
friends because of extremely-difficult-to-guarantee synchronization
between the host clock and the "hardware clock" in Qemu guests.

This is most prevalently seen when running kernel benchmarks, SMP
locking, and SMP thread stress tests when running multiple SMP Qemu
guests in parallel.

It reduces to the fact that one Qemu instance will make assumptions
about host CPU availability that break down under heavy load. That
causes jitter in the guest "hardware clock", typically on the order
of a few ms or sometimes tens of ms.

We don't (yet) run those kinds of stress tests involving multiple
Qemu targets in our CI, so retrying is really just exacerbating
a problem, which is that if our hardware platforms fail on the first try,
they typically fail on subsequent retries of the same test.

Retrying is therefore just a waste of cycles and clogs-up CI.

Let's eliminate retries until they are absolutely necessary.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>