soc: tt_blackhole: default to overwrite mode for log backend rtt
The main thing we would like to use the RTT log backend for is to ensure
we capture some logs in CI systems if PCIe ever fails to enumerate. We
can then dump those logs in the case that the CI job fails. Once PCIe
enumerates, we really don't need to care about the RTT log backend any
longer, since the virtual uart will still mostly work even if
firmware has hung. Eventually, we will likely enable dynamic log
configuration, which will allow us to programatically disable the RTT
log backend after PCIe has enumerated (from the application).

The default mode for the RTT logging backend is to _block_ which means,
that if another process on the host is not actively emptying the RTT
log buffer, it could prevent all threads on the device from executing,
depending on the log mode.

Default the log backend to overwrite, as that would mean we lose older
messages in favour of newer ones.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>