scripts: tooling: open-source blackhole_recovery tool
Open source the (previously internal) blackhole_recovery tool which is
used by CI to recovery cards after non-production or test firmware has
been flashed to them.

This tool uses the PyOCD Python API and ARM FLM / FlashAlgo functionality
to effectively use the on-board stm32 on Blackhole as a SPI flash
programmer. All of this works through the SWD protocol.

With the Python API, and some knowledge about the stm32 SoC on each
PCIe product, we are able to additionally control GPIO by reading and
writing to specific areas.

All of this code is run directly from SRAM on the stm32, so in order to
revert to production FW, all that is needed is a reset, which typically
occurs after the SWD exchange is finished.

The intent is to leverage this interface for a slightly more comprehensive
west runner that will do everything CI needs to be done with a single
invocation. This should improve CI reliability as we currently depend on a
two-phase flashing process (via twister!) as the stm32 and arc cores are
on separate JTAG adapters.

This code-drop is from the revision below in the internal repository.

4440dc45aca6760cab7f80c33835f0409aefd07b

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>