scripts: tt_bootstrap: add tt_bootstrap runner
Add a tt_bootstrap runner. This runner supports flashing a binary onto a
card that does not currently have a functional firmware, by loading the
FW using the STM32 chip for backdoor access to the SPI ROM. The binary
is loaded from a provided fwbundle.

Since firmware is not assumed to be functional, the board ID will be
reprogrammed. If one is not provided it will be calculated automatically
based on the board name.

If a hex file is provided, that will be programmed directly instead of
the firmware bundle. This method of programming only supports one ASIC
at a time, and does not recalculate the board ID.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>