ci: move metal tests to nightly temporarily
Metal tests can sometimes cause single PRs to queue for 40 minutes
while waiting for a self-hosted runner to become available.

This is the case currently for both p100a and p150a, but it is much more
pronounced for p150a since we only have one self-hosted runner with that
hardware. There are 2 self-hosted runners with p100a hardware.

Move metal tests to nightly, since the wait time is impacting
developer productivity.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>