app: dmc: increase stack space for main thread
With the recent changes to i2c / smbus, a hardfault was encountered
on line 521 of smbus_stm32.c .

```cpp
int res = i2c_transfer(config->i2c_dev, messages, num_msgs, periph_addr);
```

config is compile-time constant, non-NULL, messages is a stack variable,
and the others are in registers. Nothing out of the ordinary.

CONFIG_MAIN_STACK_SIZE=1024 though, which is not large. Bump up the
stack size so that we avoid a potential stack overflow.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>