ci: remove specific versions of pyluwen from prepare-zephyr
Previoously, a newer revision in pyluwen was required in order to work
around an issue, but there was not pypi package for it. Now that there
is a pypi package, simply switch to using that and remove the additional
command to install a specific version of pyluwen in the prepare-zephyr
workflow.

Additionally, for "Flash the firmware" in "build-fw-artifact", bump the
tt-flash version to be consistent with that which is recommended in
requirements.txt.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>