libs: tenstorrent: bh_arc: Reduce mem footprint of smbus_target
Reduce the memory overhead of SMBUS target code by using a
LUT model. When finding the CMD to execute based on the ID, the
table will provide a pointer to the command processor to use.

This allows us to only store the CMD handler data for the commands
we actually use, instead of all 256. The cost is an additional 256 ptrs.

Signed-off-by: James Growden <jgrowden@tenstorrent.com>