doc: migration: 18.8: bump pyluwen to >= 0.7.11
Our CI requires that we manually install a version of pyluwen that is
explicitly > 0.7.9. The exact revision is d20b8d6f64c0 and can be seen
here:

http://bit.ly/3HwziOj

To be sure that we communicate to end users what the required minimum
version of pyluwen is, suggest the latest version available from pypi.org
which is 0.7.11 .

Presumably, that includes additional fixes.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>