lib: bh_arc: spi_eeprom: make all local functions static
Other functions are not being used currently.

The only reasons this file still exists are

1. to populate RESET_UNIT_SCRATCH_RAM_REG_ADDR(10) with the address of
  spi_global_buffer
2. register application message handlers for MSG_TYPE_READ_EEPROM and
  MSG_TYPE_WRITE_EEPROM

The first item above is unfortunately a bit racey. We also cannot set
the value statically, since it's in scratch ram.

Ideally, the host would query the location of this buffer using a more
elaborate query / response framework (like RPC).

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>