lib: tenstorrent: bh_arc: Fix BWBR PEC
SMBUS spec indicates that, for block-write-block-read transactions,
the PEC is only present on the end of the read phase (and covers all
of the transaction.) The bug before was that we were also expecting
a PEC byte during the write phase, which is incorrect.

Add a test command to cover this case and a testcase in native_sim.

Signed-off-by: James Growden <jgrowden@tenstorrent.com>