scripts: tt_boot_fs: use 'reg' property to set image source location
Use the 'reg' property to set the image source location. Note that the
'source' property can still override this (needed for boardcfg).
Otherwise, all image locations are fixed unless they use the "$AUTO"
property.

Use the "$AUTO" property for most partitions. We will remove this
property in an overlay to make a fixed image.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>