zephyr: patches: Patch to correct i2c_dw target
i2c_dw target implementation has a bug where every byte
is processed by both CB functions, and only the bytes
from read_requested cb are transmitted on the bus.

The proper implementation is to only call the read_requested
cb on the first byte of the transaction, then call the
read_processed cb on every other byte of the transaction.
Every byte returned by the two functions should be
transmitted on the wire once, in order.

Signed-off-by: James Growden <jgrowden@tenstorrent.com>