ci: hardware-long: don't flash DMFW directly, update it with tt-flash
Rather than flashing the DMFW directly, use tt-flash to update it. This
step should be safe for the hardware-long job because the runner will
restore the card to a working firmware before we start execution.

We do need to enable SMBUS tests in the DMFW, so add that to the E2E
stress build settings twister uses.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>